Move freeze functions to their own runtime module
